+++ src/plugins/storage-dexie/rx-storage-instance-dexie.ts
@@ -6,7 +6,8 @@ import {
     now,
     ensureNotFalsy,
     defaultHashSha256,
-    RXDB_UTILS_GLOBAL
+    RXDB_UTILS_GLOBAL,
+    PREMIUM_FLAG_HASH
 } from '../utils/index.ts';
 import type {
     RxStorageInstance,
@@ -87,16 +88,16 @@ export class RxStorageInstanceDexie<RxDocType> implements RxStorageInstance<
             (
                 !RXDB_UTILS_GLOBAL.premium ||
                 typeof RXDB_UTILS_GLOBAL.premium !== 'string' ||
-                (await defaultHashSha256(RXDB_UTILS_GLOBAL.premium) !== '6da4936d1425ff3a5c44c02342c6daf791d266be3ae8479b8ec59e261df41b93')
+                (await defaultHashSha256(RXDB_UTILS_GLOBAL.premium) !== PREMIUM_FLAG_HASH)
             )
         ) {
             console.warn(
                 [
                     '-------------- RxDB Open Core RxStorage -------------------------------',
-                    'You are using the free Dexie.js based RxStorage implementation from RxDB https://rxdb.info/rx-storage-dexie.html ',
+                    'You are using the free Dexie.js based RxStorage implementation from RxDB https://rxdb.info/rx-storage-dexie.html?console=dexie ',
                     'While this is a great option, we want to let you know that there are faster storage solutions available in our premium plugins.',
                     'For professional users and production environments, we highly recommend considering these premium options to enhance performance and reliability.',
-                    ' https://rxdb.info/premium ',
+                    ' https://rxdb.info/premium?console=dexie ',
                     'If you already purchased premium access you can disable this log by calling the setPremiumFlag() function from rxdb-premium/plugins/shared.',
                     '---------------------------------------------------------------------'
                 ].join('\n')